+++ src/app/pdf_to_image.py
@@ -5,9 +5,10 @@ from tkinter import ttk, Image
 from tkinter.filedialog import askdirectory
 
 import fitz
+import tkinterdnd2
 from PIL import Image
 
-from utility import get_treeview_file_list
+from utility import get_treeview_file_list, drop_pdf_files_to_treeview
 from widget import Process, FrameTitle, OutputFolder, FileList
 
 
@@ -34,6 +35,12 @@ class PDFToImage(ttk.Frame):
         self.Process.configure(text=_('PDF to Image'))
         self.Process.ButtonProcess.configure(text=_('Convert'), command=self.pdf_to_image)
         self.Process.pack(fill='x', padx=4, pady=4)
+
+        self.drop_target_register(tkinterdnd2.DND_FILES)
+        self.dnd_bind('<<Drop>>', self.drop_files)
+
+    def drop_files(self, event):
+        drop_pdf_files_to_treeview(self.Filelist.TreeviewFilelist, event)
 
     def set_output_folder(self):
         initial_dir = None
@@ -141,7 +148,7 @@ class Options(ttk.LabelFrame):
 
 
 if __name__ == '__main__':
-    root = tk.Tk()
+    root = tkinterdnd2.Tk()
     root.title('PDF to Image')
     pdf_to_image = PDFToImage(root)
     pdf_to_image.pack(expand=True, fill='both', padx=4, pady=4)